(#2101) Allow report summary to not be shown
As part of a normal Cake execution, a report summary is always shown at
the end. This includes information about what tasks were ran, how long
they took, etc.  However, there are those that don't want to allow this
to happen.  This commit introduces a new configuration option, which
allows the report summary to not be shown. This can be done in the
usual way with a command line parameter, cake.config file entry, or an
environment variable.